Fixes critical errors.
